refactor: move into helper function file
