+++ src/app/(core)/teachers/programmes/[slug]/[tab]/page.test.tsx
@@ -2,7 +2,7 @@
  * @jest-environment node
  */
 import { getProgrammeData, getSubjectPhaseOptions } from "./getProgrammeData";
-import ProgrammePageTabs, { generateMetadata, getFileSizes } from "./page";
+import ProgrammePageTabs, { generateMetadata } from "./page";
 
 import getBrowserConfig from "@/browser-lib/getBrowserConfig";
 import { resolveOakHref } from "@/common-lib/urls";
@@ -11,7 +11,6 @@ import { createUnit } from "@/fixtures/curriculum/unit";
 import { curriculumOverviewMVFixture } from "@/node-lib/curriculum-api-2023/fixtures/curriculumOverview.fixture";
 import curriculumPhaseOptionsFixture from "@/node-lib/curriculum-api-2023/fixtures/curriculumPhaseOptions.fixture";
 import { filterValidCurriculumPhaseOptions } from "@/pages-helpers/curriculum/docx/tab-helpers";
-import { contentLengthFromResource } from "@/utils/resource";
 
 const defaultParams = new URLSearchParams("");
 const mockUseSearchParams = jest.fn(() => defaultParams);
@@ -35,10 +34,6 @@ jest.mock("next/navigation", () => {
     },
   };
 });
-
-jest.mock("@/utils/resource", () => ({
-  contentLengthFromResource: jest.fn().mockResolvedValue(1234),
-}));
 
 jest.mock("next/headers", () => ({
   draftMode: jest.fn().mockResolvedValue({
@@ -278,131 +273,6 @@ describe("Programme page tabs", () => {
         searchParams: Promise.resolve({}),
       }),
     ).rejects.toEqual(new Error("NEXT_HTTP_ERROR_FALLBACK;404"));
-  });
-});
-
-describe("getFileSizes", () => {
-  test("expect download size to be in document (tiers)", async () => {
-    (contentLengthFromResource as jest.Mock).mockImplementation(
-      async (url: string) => {
-        if (url.includes("foundation")) {
-          return 123456;
-        } else if (url.includes("higher")) {
-          return 223456;
-        }
-        return -1;
-      },
-    );
-    const out = await getFileSizes(
-      {
-        phaseSlug: "",
-        subjectSlug: "",
-        ks4OptionSlug: null,
-      },
-      {
-        tiers: [
-          {
-            tier: "Foundation",
-            tier_slug: "foundation",
-          },
-          {
-            tier: "Higher",
-            tier_slug: "higher",
-          },
-        ],
-        child_subjects: [],
-      },
-      1,
-    );
-
-    expect(out).toEqual([
-      {
-        downloadId: "curriculum-plans",
-        size: 123456,
-        tier: "foundation",
-        childSubject: null,
-      },
-      {
-        downloadId: "curriculum-plans",
-        size: 223456,
-        tier: "higher",
-        childSubject: null,
-      },
-    ]);
-  });
-
-  test("expect download size to be in document (child subjects & tiers)", async () => {
-    (contentLengthFromResource as jest.Mock).mockImplementation(
-      async (url: string) => {
-        if (url.includes("foundation") && url.includes("combined-science")) {
-          return 123456;
-        } else if (url.includes("higher") && url.includes("combined-science")) {
-          return 223456;
-        } else if (url.includes("foundation") && url.includes("biology")) {
-          return 323456;
-        } else if (url.includes("higher") && url.includes("biology")) {
-          return 423456;
-        }
-        return -1;
-      },
-    );
-    const out = await getFileSizes(
-      {
-        phaseSlug: "",
-        subjectSlug: "",
-        ks4OptionSlug: null,
-      },
-      {
-        tiers: [
-          {
-            tier: "Foundation",
-            tier_slug: "foundation",
-          },
-          {
-            tier: "Higher",
-            tier_slug: "higher",
-          },
-        ],
-        child_subjects: [
-          {
-            subject: "Combined science",
-            subject_slug: "combined-science",
-          },
-          {
-            subject: "Biology",
-            subject_slug: "biology",
-          },
-        ],
-      },
-      1,
-    );
-
-    expect(out).toEqual([
-      {
-        downloadId: "curriculum-plans",
-        size: 123456,
-        tier: "foundation",
-        childSubject: "combined-science",
-      },
-      {
-        downloadId: "curriculum-plans",
-        size: 223456,
-        tier: "higher",
-        childSubject: "biology",
-      },
-      {
-        downloadId: "curriculum-plans",
-        size: 323456,
-        tier: "foundation",
-        childSubject: "combined-science",
-      },
-      {
-        downloadId: "curriculum-plans",
-        size: 423456,
-        tier: "higher",
-        childSubject: "biology",
-      },
-    ]);
   });
 });
